+++ CHANGES.rst
@@ -21,6 +21,14 @@ Changes
 =======
 
 
+Version 0.8.4
+-------------
+
+- Add support to generate data cube from Landsat Collection 2 (`#172 <https://github.com/brazil-data-cube/cube-builder/issues/172>`_)
+- Add support to combine Landsat Collection 2 sensors (L5/L7, L7/L8, L7/L8/L9) using single collection (`#172 <https://github.com/brazil-data-cube/cube-builder/issues/172>`_)
+- Review API error when no parameter is set.
+
+
 Version 0.8.3 (2022-10-03)
 --------------------------